pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cerebrofy"
version = "1.0.4"
description = "AI-powered codebase intelligence CLI"
readme = "README.md"
Home-page = "https://github.com/mm0rsy/Cerebrofy"
license = { file = "LICENSE" }
authors = [{ name = "Mohamed Morsy"}]
requires-python = ">=3.11"
dependencies = [
    "tree-sitter>=0.21,<0.22",
    "tree-sitter-languages>=1.10,<2",
    "pathspec>=0.12",
    "PyYAML>=6.0",
    "click>=8.1",
    "sqlite-vec>=0.1.6",
    "rich-click>=1.9.7",
    "fastembed>=0.3",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0"]

[project.scripts]
cerebrofy = "cerebrofy.cli:main"

[dependency-groups]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1",
    "types-PyYAML>=6.0",
]

[tool.pytest.ini_options]
addopts = "--cov=cerebrofy --cov-report=term-missing --cov-report=xml --cov-fail-under=70"
testpaths = ["tests"]

[tool.coverage.run]
source = ["cerebrofy"]
omit = ["*/mcp/server.py"]  # optional MCP server — heavy I/O, covered by integration tests

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "\\.\\.\\.",
]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/cerebrofy"]

[tool.hatch.build.targets.wheel.force-include]
"src/cerebrofy/queries" = "cerebrofy/queries"
"src/cerebrofy/skills" = "cerebrofy/skills"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true

# mcp is an optional dependency that ships no type stubs.
# Suppress import-not-found for the whole mcp namespace.
[[tool.mypy.overrides]]
module = ["mcp", "mcp.*"]
ignore_missing_imports = true

# mcp decorators (list_tools, call_tool) raise either no-untyped-call or
# untyped-decorator depending on mypy's evaluation order — never both at once.
# Disable warn_unused_ignores locally so both codes can coexist safely.
[[tool.mypy.overrides]]
module = ["cerebrofy.mcp.server"]
warn_unused_ignores = false
